requirements.txt
django>=3.2.14
mysqlclient
djangorestframework
django-cors-headers
pyyaml
uritemplate